Now find ipopt using its .pc file.
Also, add a patch: since the include path in ipopt.pc has the following
form "Path/include/coin", a post-treatment process is added to have
instead "Path/include", leaving the header inclusion untouched.
Also, this include path is added to the file roboptim-core-plugin-ipopt.pc,
to ensure the compilation of the projects depending on this plugin.